requirements.txt
fastapi
uvicorn
python-dotenv
zstandard==0.24.0
requests
openai
faiss-cpu
numpy
elevenlabs
